add: parking space model change: change types from ParkingSpace to IParkingSpace for consistency
