+++ client/src/routes/editor/utils/index.ts
@@ -9,21 +9,3 @@ export const calculateCenterPosition = (height: number, width: number) => {
 
   return { x: centerX, z: centerZ}  
 }
-
-function hexToRGB(hex: string) {
-  const result = /^#?([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})$/i.exec(hex);
-  return result ? {
-    r: parseInt(result[1], 16),
-    g: parseInt(result[2], 16),
-    b: parseInt(result[3], 16)
-  } : null;
-}
-
-export function selectMatchingContrastColor(hex: string, darkColor = "#000000", lightColor = "#ffffff") {
-  const rgb = hexToRGB(hex);
-  if ( !rgb ) return null;
-
-  const { r, g, b } = rgb;
-  const yiq = (r * 299 + g * 587 + b * 114) / 1000;
-  return (yiq >= 128) ? darkColor : lightColor;
-}
